partial completion of comments from the mentor
